requirements.txt
matplotlib>=3.3.0
numpy>=1.22.2
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0
psutil
py-cpuinfo
thop>=0.1.1
pandas>=1.1.4
seaborn>=0.11.0

# Optional dependencies
# ipython
# check-manifest
# pre-commit
# pytest
# pytest-cov
# coverage[toml]
# mkdocs-material
# mkdocstrings[python]
# mkdocs-jupyter
# mkdocs-redirects
# mkdocs-ultralytics-plugin>=0.0.43
# lancedb
# duckdb
# streamlit
# onnx>=1.12.0
# coremltools>=7.0; platform_system != 'Windows'
# openvino-dev>=2023.0
# tensorflow<=2.13.1
# tensorflowjs>=3.9.0
# hub-sdk>=0.0.2
# albumentations>=1.0.3
# pycocotools>=2.0.6
# comet
# tensorboard>=2.13.0
# dvclive>=2.12.0
